+++ Fluid-HTNCPP.UnitTests/SequenceTests.cpp
@@ -2,7 +2,7 @@
 #include "pch.h"
 #include "CppUnitTest.h"
 #include "Contexts/BaseContext.h"
-#include "Domain.h"
+#include "CoreIncludes/Domain.h"
 #include "Planners/Planner.h"
 #include "Tasks/CompoundTasks/Sequence.h"
 #include "Tasks/CompoundTasks/Selector.h"
