+++ Fluid-HTNCPP/Fluid-HTNCPP.vcxproj
@@ -175,18 +175,19 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="BaseDomainBuilder.h" />
     <ClInclude Include="Conditions\Condition.h" />
     <ClInclude Include="Contexts\BaseContext.h" />
     <ClInclude Include="Contexts\Context.h" />
+    <ClInclude Include="CoreIncludes\BaseDomainBuilder.h" />
+    <ClInclude Include="CoreIncludes\Domain.h" />
+    <ClInclude Include="CoreIncludes\pch.h" />
+    <ClInclude Include="CoreIncludes\STLTypes.h" />
+    <ClInclude Include="CoreIncludes\WorldState.h" />
     <ClInclude Include="DebugInterfaces\DecompositionLogEntry.h" />
-    <ClInclude Include="Domain.h" />
     <ClInclude Include="Effects\Effect.h" />
     <ClInclude Include="Effects\EffectType.h" />
     <ClInclude Include="Operators\Operator.h" />
-    <ClInclude Include="pch.h" />
     <ClInclude Include="Planners\Planner.h" />
-    <ClInclude Include="STLTypes.h" />
     <ClInclude Include="Tasks\CompoundTasks\CompoundTask.h" />
     <ClInclude Include="Tasks\CompoundTasks\DecompositionStatus.h" />
     <ClInclude Include="Tasks\CompoundTasks\PausePlanTask.h" />
@@ -196,16 +197,10 @@
     <ClInclude Include="Tasks\OtherTasks\Slot.h" />
     <ClInclude Include="Tasks\PrimitiveTasks\PrimitiveTask.h" />
     <ClInclude Include="Tasks\Task.h" />
-    <ClInclude Include="WorldState.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Effects\Effect.cpp" />
-    <ClCompile Include="pch.cpp">
-      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
-      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
-      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
-      <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
-    </ClCompile>
+    <ClCompile Include="pch.cpp" />
     <ClCompile Include="Tasks\CompoundTasks\CompoundTask.cpp" />
     <ClCompile Include="Tasks\CompoundTasks\Selector.cpp" />
     <ClCompile Include="Tasks\CompoundTasks\Sequence.cpp" />
